# Add support for expanded /hosts endpoint output