Adds support for the expanded /hosts endpoint output introduced in unioslo/mreg#572. Integration tests were performed against the container image produced here locally: https://github.com/unioslo/mreg/actions/runs/13944068027

The test suite will not pass here until we have merged unioslo/mreg#572.