@terjekv 2025-03-20
Just out of curiosity and as an answer to "why", can you compare the runtimes of `host info` for a host with and without the changes?