@pederhan 2025-03-20
> Just out of curiosity and as an answer to "why", can you compare the runtimes of `host info` for a host with and without the changes?

Take this with a grain of salt:

Benchmarked with hyperfine using:

```
hyperfine --warmup 5 -n 20 "mreg-cli --url https://<redacted> --domain uio.no --user <redacted> host info testhost"
```

On this branch (36a8cb8)

```
Benchmark 1: 36a8cb8
  Time (mean ± σ):     907.8 ms ± 177.7 ms    [User: 279.7 ms, System: 49.3 ms]
  Range (min … max):   777.6 ms … 1516.9 ms    20 runs
```
  
  
On master:
  
```
Benchmark 1: master
  Time (mean ± σ):      1.393 s ±  0.052 s    [User: 0.290 s, System: 0.051 s]
  Range (min … max):    1.322 s …  1.508 s    20 runs
```