@terjekv 2025-03-20
It'd be interesting to see the start + logout, but it looks like a big improvement. Good stuffs, thank you!